refactor: persistent-mode 코어를 lib/persistent-mode-core로 추출
decision·state·transcript-detector·types·utils(+테스트) 9개 파일을
hooks/persistent-mode/에서 lib/persistent-mode-core/로 git mv.
move-only 리팩터, 로직 무변경(makeDecision 본문 diff 0). 남은 Claude
어댑터(index.ts, stdin.ts, stdin.test.ts) import 경로 4줄만
@lib/persistent-mode-core로 재작성.

Codex 어댑터가 동일 코어를 공유할 수 있도록 하기 위한 선행 작업.
