+++ lib/persistent-mode-core/decision.test.ts
