+++ lib/persistent-mode-core/decision.ts
