+++ lib/persistent-mode-core/state.test.ts
