+++ lib/persistent-mode-core/state.ts
