+++ lib/persistent-mode-core/transcript-detector.test.ts
