+++ lib/persistent-mode-core/transcript-detector.ts
