+++ lib/persistent-mode-core/utils.test.ts
