+++ lib/persistent-mode-core/utils.ts
